fix(chat): forced full reload on code update
