+++ src/components/BouncingDots/Spinner.module.scss
