+++ src/modules/chat/assistant-plan/PlanStep.tsx
@@ -24,7 +24,7 @@ import { encodeEntityWithMetadata } from '@/app/api/utils';
 import { ExpandPanel } from '@/components/ExpandPanel/ExpandPanel';
 import { ExpandPanelButton } from '@/components/ExpandPanelButton/ExpandPanelButton';
 import { LineClampText } from '@/components/LineClampText/LineClampText';
-import { Spinner } from '@/components/Spinner/Spinner';
+import { Spinner } from '@/components/BouncingDots/Spinner';
 import { Tooltip } from '@/components/Tooltip/Tooltip';
 import { useToolInfo } from '@/modules/tools/hooks/useToolInfo';
 import { fadeProps } from '@/utils/fadeProps';
@@ -43,7 +43,6 @@ import { AnimatePresence, motion } from 'framer-motion';
 import JSON5 from 'json5';
 import { ReactElement, useCallback, useEffect, useId, useMemo } from 'react';
 import { useThreadApi } from '../hooks/useThreadApi';
-import { useChat } from '../providers/ChatProvider';
 import {
   useExpandedStep,
   useExpandedStepActions,
@@ -57,6 +56,7 @@ import classes from './PlanStep.module.scss';
 import { useUserSetting } from '@/layout/hooks/useUserSetting';
 import { getToolApproval, getToolReferenceFromToolCall } from './utils';
 import { useAppContext } from '@/layout/providers/AppProvider';
+import { useChat } from '../providers/chat-context';
 
 interface Props {
   step: AssistantPlanStep;
