+++ src/modules/chat/assistant-plan/PlanWithSources.tsx
@@ -24,7 +24,7 @@ import clsx from 'clsx';
 import { useEffect, useMemo, useState } from 'react';
 import { v4 as uuid } from 'uuid';
 import { SourcesView } from '../layout/SourcesView';
-import { useChat } from '../providers/ChatProvider';
+import { useChat } from '../providers/chat-context';
 import {
   ExpandedStepProvider,
   useExpandedStep,
@@ -43,7 +43,7 @@ import { TraceData } from '../trace/types';
 import { useBuildTraceData } from '../trace/useBuildTraceData';
 import { TraceInfoView } from '../trace/TraceInfoView';
 import { TraceDataProvider } from '../trace/TraceDataProvider';
-import { Spinner } from '@/components/Spinner/Spinner';
+import { Spinner } from '@/components/BouncingDots/Spinner';
 import { MAX_API_FETCH_LIMIT } from '@/app/api/utils';
 import { useAppContext } from '@/layout/providers/AppProvider';
