+++ src/modules/chat/message/MessageLoading.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Spinner } from '@/components/Spinner/Spinner';
+import { Spinner } from '@/components/BouncingDots/Spinner';
 import classes from './MessageLoading.module.scss';
 
 interface Props {
