+++ src/modules/chat/message/markdown/AttachmentImage.tsx
@@ -16,7 +16,7 @@
 
 import { useQuery } from '@tanstack/react-query';
 import classes from './AttachmentImage.module.scss';
-import { Spinner } from '@/components/Spinner/Spinner';
+import { Spinner } from '@/components/BouncingDots/Spinner';
 import { readFileContent } from '@/app/api/files';
 import { useAppContext } from '@/layout/providers/AppProvider';
