+++ src/modules/chat/types.ts
@@ -22,6 +22,7 @@ import { Assistant } from '../assistants/types';
 import { AssistantPlan, ThreadRun } from '@/app/api/threads-runs/types';
 import { VectoreStoreFileUpload } from '../knowledge/files/VectorStoreFilesUploadProvider';
 import { EntityWithDecodedMetadata } from '@/app/api/types';
+import { Thread } from '@/app/api/threads/types';
 
 export interface ThreadAssistant {
   name?: string;
